LightGBM wrapping API
We begin by creating a low-level wrapper around the command line
lightgbm executable.  Note that We use some refined types to get a bit
more precision in the parameter types.

We then provide a higher-level API built on the low-level wrapper.

We also get rid of the useless application in the package.yaml - this
is a library not an app.